trained model and built streamlit application

+++ books_recommender/components/stage_00_data_ingestion.py
@@ -23,23 +23,29 @@ class DataIngestion:
     
     def download_data(self):
         """
-        Fetch the data from the url
-        
+        Fetch the data from the URL if not already present.
         """
         try:
-            
             dataset_url = self.data_ingestion_config.dataset_download_url
             zip_download_dir = self.data_ingestion_config.raw_data_dir
             os.makedirs(zip_download_dir, exist_ok=True)
+
             data_file_name = os.path.basename(dataset_url)
             zip_file_path = os.path.join(zip_download_dir, data_file_name)
-            logging.info(f"Downloading data from {dataset_url} into file {zip_file_path}")
-            urllib.request.urlretrieve(dataset_url,zip_file_path)
-            logging.info(f"Downloaded data from {dataset_url} into file {zip_file_path}")
+
+            # Avoid re-downloading if file already exists
+            if not os.path.exists(zip_file_path):
+                logging.info(f"Downloading data from {dataset_url} into file {zip_file_path}")
+                urllib.request.urlretrieve(dataset_url, zip_file_path)
+                logging.info(f"Downloaded data into file {zip_file_path}")
+            else:
+                logging.info(f"Download skipped. File already exists at: {zip_file_path}")
+
             return zip_file_path
 
         except Exception as e:
             raise AppException(e, sys) from e
+
 
 
     def extract_zip_file(self,zip_file_path: str):
